+++ src/app/admin/login/page.tsx
@@ -1,65 +1,10 @@
-"use client";
+import type { Metadata } from "next";
+import { LoginForm } from "@/components/admin/login-form";
 
-import { useState } from "react";
-import { useRouter } from "next/navigation";
+export const metadata: Metadata = {
+  title: "管理画面ログイン",
+};
 
 export default function AdminLoginPage() {
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState("");
-  const [loading, setLoading] = useState(false);
-  const router = useRouter();
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setError("");
-    setLoading(true);
-
-    try {
-      const res = await fetch("/api/admin/login", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify({ password }),
-      });
-
-      if (res.ok) {
-        router.push("/admin/orders");
-      } else {
-        const data = await res.json();
-        setError(data.error || "ログインに失敗しました");
-      }
-    } catch {
-      setError("通信エラーが発生しました");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="flex min-h-screen items-center justify-center bg-gray-50">
-      <form
-        onSubmit={handleSubmit}
-        className="w-full max-w-sm space-y-4 rounded-lg bg-white p-8 shadow"
-      >
-        <h1 className="text-xl font-bold">管理画面ログイン</h1>
-        {error && (
-          <p className="text-sm text-red-600">{error}</p>
-        )}
-        <input
-          type="password"
-          value={password}
-          onChange={(e) => setPassword(e.target.value)}
-          placeholder="パスワード"
-          className="w-full rounded border p-2"
-          disabled={loading}
-        />
-        <button
-          type="submit"
-          disabled={loading}
-          className="w-full rounded bg-gray-800 py-2 text-white hover:bg-gray-900 disabled:opacity-50"
-        >
-          {loading ? "ログイン中..." : "ログイン"}
-        </button>
-      </form>
-    </div>
-  );
+  return <LoginForm />;
 }
